download report pdf by certificate
